foglet-protocol fully rebuild with new API
