+++ src/fprotocol/foglet-protocol.js
@@ -24,34 +24,30 @@ SOFTWARE.
 'use strict';
 
 const AnswerQueue = require('./answer-queue.js');
-// const HandlerManager = require('./handler-manager.js');
 
 /**
  * FogletProtocol represent an abstract protocol.
  * A Protocol is a a set of behaviours used to interact with others foglet that shares the same protocol.
- *
- * One can define a new protocol by subclassing {@link FogletProtocol} and implementing the following methods:
- * * `unicast` to define the unicast services offered by the protocol.
- * * `broadcast` to define the broadcast services offered by the protocol.
- * Once these methods have been implemented, you need to implement a `handler method` per service offered.
- * A handler is a method of the subclass where its name is the name of the related service prefixed with `_`.
- * For example, a service **get** requires to define a method named **_get** in the subclass.
  * @abstract
  * @author Thomas Minier
  * @example
- * class SimpleUnicastProtocol extends FogletProtocol {
-
-   _unicast () {
-     return [ 'get' ];
-   }
-
-   _get (msg, reply, reject) {
-     if (msg.number % 2 === 0)
-        reply('you send an event number');
-     else
-        reject('you send an event number, it is not good!!!');
-   }
- }
+ * const ExampleUnicastProtocol = defineProtocol('example-unicast-protocol')`
+ *  init
+ *  ${(self, callback, done) => {
+ *    self._callback = callback;
+ *    self._done = done;
+ *  }}
+ *  get
+ *  ${service => {
+ *    service.is.unicast();
+ *    service.on.receive(function (id, msg, reply, reject) {
+ *      if (this._callback) this._callback(id, msg, reply, reject);
+ *      if (this._done) this._done();
+ *    });
+ *  }}
+ *  `;
+ *
+ * module.exports = ExampleUnicastProtocol;
  */
 class FogletProtocol {
   /**
@@ -63,17 +59,31 @@ class FogletProtocol {
   constructor (name, foglet, ...args) {
     this._name = name;
     this._foglet = foglet;
-    // this._handlerManager = new HandlerManager(this);
     this._answerQueue = new AnswerQueue();
     this._initHandlers();
     if ('_init' in this)
       this._init(this, ...args);
   }
 
+  /**
+   * Helper to send a unicast message
+   * @private
+   * @param  {string} id  - ID of the peer to which the message should be sent
+   * @param  {*} msg  - The message to send
+   * @param  {function} resolve - Function used to resolve a related promise when an answer to the message is received
+   * @param  {function} reject  - Function used to reject a related promise when an answer to the message is received
+   * @return {void}
+   */
   _sendUnicast (id, msg, resolve, reject) {
     this._foglet.sendUnicast(this._answerQueue.stamp(msg, resolve, reject), id);
   }
 
+  /**
+   * Helper to send a broadcast message
+   * @private
+   * @param  {*} msg  - The message to send
+   * @return {void}
+   */
   _sendBroadcast (msg) {
     this._foglet.sendBroadcast(msg);
   }
@@ -84,7 +94,7 @@ class FogletProtocol {
    * @param {Object} msg - Answer received
    * @return{void}
    */
-  _answerReply (msg) {
+  _answerReply (id, msg) {
     this._answerQueue.resolve(msg.answerID, msg.value);
   }
 
@@ -94,13 +104,18 @@ class FogletProtocol {
    * @param {Object} msg - Answer received
    * @return{void}
    */
-  _answerReject (msg) {
+  _answerReject (id, msg) {
     this._answerQueue.reject(msg.answerID, msg.value);
   }
 
+  /**
+   * Initialize the reception of messages from unicast & broadcast channels
+   * @private
+   * @return {void}
+   */
   _initHandlers () {
     this._foglet.onUnicast((id, msg) => this._handleUnicast(id, msg));
-    this._foglet.onBroadcast((msg, id) => this._handleBroadcast(msg, id));
+    this._foglet.onBroadcast((msg, id) => this._handleBroadcast(id, msg));
   }
 
   /**
@@ -117,7 +132,7 @@ class FogletProtocol {
       if (msg.method !== 'answerReply' || msg.method !== 'answerReject') {
         const reply = value => {
           this._foglet.sendUnicast({
-            protocol: 'foglet-protocol',
+            protocol: this._name,
             method: 'answerReply',
             payload: {
               answerID: msg.answerID,
@@ -127,7 +142,7 @@ class FogletProtocol {
         };
         const reject = value => {
           this._foglet.sendUnicast({
-            protocol: 'foglet-protocol',
+            protocol: this._name,
             method: 'answerReject',
             payload: {
               answerID: msg.answerID,
@@ -145,67 +160,15 @@ class FogletProtocol {
   /**
    * Handle the reception of a broadcast message
    * @private
-   * @param {Object} msg - The message received
    * @param {string} senderID - ID of the peer who send the message
+   * @param {Object} msg - The message received
    * @return {void}
    */
-  _handleBroadcast (msg, senderID) {
+  _handleBroadcast (senderID, msg) {
     const handlerName = `_${msg.method}`;
     if (this._name === msg.protocol && handlerName in this) {
       this[handlerName].call(this, senderID, msg.payload);
     }
-  }
-
-  /**
-   * Build protocol services
-   * @private
-   * @deprecated
-   * @return {void}
-   */
-  _buildProtocol () {
-    // register unicast services
-    this._unicast().forEach(method => {
-      this._registerService(method, (id, payload) => {
-        return new Promise((resolve, reject) => {
-          const msg = {
-            protocol: this._name,
-            method,
-            payload
-          };
-          this._foglet.sendUnicast(this._answerQueue.stamp(msg, resolve, reject), id);
-        });
-      });
-      this._handlerManager.registerUnicastHandler(`${this._name}/${method}`, `_${method}`);
-    });
-
-    // do the same for broadcast service
-    this._broadcast().forEach(method => {
-      this._registerService(method, payload => {
-        return new Promise((resolve, reject) => {
-          const msg = {
-            protocol: this._name,
-            method,
-            payload
-          };
-          this._foglet.sendBroadcast(this._answerQueue.stamp(msg, resolve, reject));
-        });
-      });
-      this._handlerManager.registerBroadcastHandler(`${this._name}/${method}`, `_${method}`);
-    });
-  }
-
-  /**
-   * Register a service
-   * @private
-   * @deprecated
-   * @param  {string} serviceName   - The service name
-   * @param  {function} serviceMethod - The method called by the service
-   * @return {void}
-   */
-  _registerService (serviceName, serviceMethod) {
-    if (serviceName in this)
-      throw new SyntaxError(`Cannot create a new service for protocol ${this._name} if the method ${serviceName} is already defined`);
-    this[serviceName] = serviceMethod;
   }
 }
